+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Beta 0.2 (tba)
+## Beta 0.2 (08.02.2020)
 
 ### Added
 
@@ -9,6 +9,14 @@
   - Added changelog
   - Added "Reset Mop Position" button
 - Urinating button can now be binded under any key (as long as it's one of the letters, or digit)
+
+### Changes
+
+- Stain removal rate is now a constant value
+
+### Bug Fixes
+
+- Improved in hand rotation and holding
 
 ## Beta 0.1 (07.02.2020)